# Conflicts:
#	build/classes/java/main/com/sparta/newsfeed/user/repository/UserRepository.class
#	build/classes/java/main/com/sparta/newsfeed/user/service/UserService.class
#	build/tmp/compileJava/previous-compilation-data.bin

+++ src/main/java/com/sparta/newsfeed/user/service/UserService.java
@@ -3,10 +3,7 @@ package com.sparta.newsfeed.user.service;
 import com.sparta.newsfeed.auth.service.AuthService;
 import com.sparta.newsfeed.config.PasswordEncoder;
 import com.sparta.newsfeed.newsfeed.repository.NewsfeedRepository;
-import com.sparta.newsfeed.user.dto.SignUpRequestDto;
-import com.sparta.newsfeed.user.dto.SignUpResponseDto;
-import com.sparta.newsfeed.user.dto.UserRequestDto;
-import com.sparta.newsfeed.user.dto.UserResponseDto;
+import com.sparta.newsfeed.user.dto.*;
 import com.sparta.newsfeed.user.entity.User;
 import com.sparta.newsfeed.user.repository.UserRepository;
 import jakarta.servlet.http.HttpSession;
@@ -48,14 +45,10 @@ public class UserService {
     }
 
     public UserResponseDto updateUser(Long id, UserRequestDto requestDto, HttpSession session){
+
         //TODO: 에러 처리
         User foundUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id));
-
         String sessionEmail = authService.getSessionEmail(session);
-
-        if (requestDto.getOldPassword().equals(requestDto.getNewPassword())) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "You can not change with same password.");
-        }
 
         if (!foundUser.getEmail().equals(sessionEmail)) {
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You can not update other user's information.");
@@ -66,22 +59,34 @@ public class UserService {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Password is wrong.");
         }
 
+        if (requestDto.getOldPassword().equals(requestDto.getNewPassword())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "You can not change with same password.");
+        }
+
         foundUser.partialUpdate(
                 requestDto.getName(),
-                requestDto.getEmail(),
                 passwordEncoder.encode(requestDto.getNewPassword())
         );
 
         return new UserResponseDto(id, requestDto.getEmail(), requestDto.getName());
     }
 
+    public void deleteUser(Long id, UserDeleteRequestDto requestDto , HttpSession session){
+        User foundUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id));
+        String email = authService.getSessionEmail(session);
 
-    public void deleteUser(Long id, HttpSession session){
+        if (!email.equals(foundUser.getEmail())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "You can not delete other user's account.");
+        }
+
+        //TODO:겹치는 코드 , 메서드 추출 ?
+        if (!passwordEncoder.matches(requestDto.getPassword(), foundUser.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Password is wrong.");
+        }
+
         newsfeedRepository.deleteNewsfeedsByUserId(id);
         //TODO: 에러 처리
-        User foundUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id));
         foundUser.updateSoftDelete();
-
         authService.logout(session);
     }
